+++ packages/components/Avatar/src/AvatarTokens.ts
@@ -10,7 +10,7 @@ export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = (t: Theme
     avatarOpacity: 1,
     fontFamily: t.typography.families.primary,
     fontWeight: globalTokens.font.weight.semibold,
-    fontSize: globalTokens.font.size[200],
+    fontSize: globalTokens.font.size200,
     size: 24,
     iconSize: 16,
     iconColor: t.colors.neutralForeground3,
@@ -18,10 +18,10 @@ export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = (t: Theme
     borderColor: t.colors.neutralStroke1,
     borderWidth: t.host.appearance === 'highContrast' ? 1 : 0,
     circular: {
-      borderRadius: globalTokens.corner.radius.circle,
+      borderRadius: globalTokens.corner.radiusCircular,
     },
     square: {
-      borderRadius: globalTokens.corner.radius.medium,
+      borderRadius: globalTokens.corner.radius40,
     },
     inactive: {
       avatarOpacity: 0.8,
@@ -30,18 +30,18 @@ export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = (t: Theme
       size: 20,
       badgeSize: 'tiny',
       iconSize: 16,
-      fontSize: globalTokens.font.size[100],
+      fontSize: globalTokens.font.size100,
       square: {
-        borderRadius: globalTokens.corner.radius.small,
+        borderRadius: globalTokens.corner.radius20,
       },
     },
     size24: {
       size: 24,
       badgeSize: 'tiny',
       iconSize: 16,
-      fontSize: globalTokens.font.size[100],
+      fontSize: globalTokens.font.size100,
       square: {
-        borderRadius: globalTokens.corner.radius.small,
+        borderRadius: globalTokens.corner.radius20,
       },
     },
     size28: {
@@ -49,13 +49,13 @@ export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = (t: Theme
       badgeSize: 'extraSmall',
       iconSize: 20,
       fontWeight: globalTokens.font.weight.semibold,
-      fontSize: globalTokens.font.size[100],
+      fontSize: globalTokens.font.size100,
     },
     size32: {
       size: 32,
       badgeSize: 'extraSmall',
       iconSize: 20,
-      fontSize: globalTokens.font.size[100],
+      fontSize: globalTokens.font.size100,
     },
     size36: {
       size: 36,
@@ -76,27 +76,27 @@ export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = (t: Theme
       size: 56,
       badgeSize: 'medium',
       iconSize: 28,
-      fontSize: globalTokens.font.size[400],
+      fontSize: globalTokens.font.size400,
       square: {
-        borderRadius: globalTokens.corner.radius.large,
+        borderRadius: globalTokens.corner.radius60,
       },
     },
     size64: {
       size: 64,
       badgeSize: 'large',
       iconSize: 32,
-      fontSize: globalTokens.font.size[500],
+      fontSize: globalTokens.font.size500,
       square: {
-        borderRadius: globalTokens.corner.radius.large,
+        borderRadius: globalTokens.corner.radius60,
       },
     },
     size72: {
       size: 72,
       badgeSize: 'large',
       iconSize: 32,
-      fontSize: globalTokens.font.size[500],
+      fontSize: globalTokens.font.size500,
       square: {
-        borderRadius: globalTokens.corner.radius.large,
+        borderRadius: globalTokens.corner.radius60,
       },
     },
     size96: {
@@ -104,9 +104,9 @@ export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = (t: Theme
       badgeSize: 'extraLarge',
       iconSize: 48,
       fontWeight: globalTokens.font.weight.regular,
-      fontSize: globalTokens.font.size[700],
+      fontSize: globalTokens.font.size700,
       square: {
-        borderRadius: globalTokens.corner.radius.extraLarge,
+        borderRadius: globalTokens.corner.radius80,
       },
     },
     size120: {
@@ -114,9 +114,9 @@ export const defaultAvatarTokens: TokenSettings<AvatarTokens, Theme> = (t: Theme
       badgeSize: 'extraLarge',
       iconSize: 48,
       fontWeight: globalTokens.font.weight.regular,
-      fontSize: globalTokens.font.size[900],
+      fontSize: globalTokens.font.size900,
       square: {
-        borderRadius: globalTokens.corner.radius.extraLarge,
+        borderRadius: globalTokens.corner.radius80,
       },
     },
     neutral: {