+++ packages/components/Badge/src/PresenceBadge/PresenceBadgeTokens.win32.ts
@@ -50,7 +50,7 @@ export const defaultPresenceBadgeTokens: TokenSettings<PresenceBadgeTokens> = (t
     unknown: getBadgeColor('red', t),
     doNotDisturb: getBadgeColor('red', t),
     offline: {
-      iconColor: isHighContrast(t) ? t.colors.neutralForeground3 : globalTokens.color.grey[38],
+      iconColor: isHighContrast(t) ? t.colors.neutralForeground3 : globalTokens.color.grey38,
     },
     outOfOffice: getBadgeColor('berry', t),
   } as PresenceBadgeTokens);