+++ packages/components/Checkbox/CHANGELOG.md
@@ -1,8 +1,45 @@
 # Change Log - @fluentui-react-native/checkbox
 
-This log was last generated on Wed, 09 Nov 2022 22:52:11 GMT and should not be manually modified.
+This log was last generated on Wed, 16 Nov 2022 08:47:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.4
+
+Wed, 16 Nov 2022 08:47:25 GMT
+
+### Patches
+
+- add initial checkbox tokens (prasadro@microsoft.com)
+- Bump @fluentui-react-native/framework to v0.8.20
+- Bump @fluentui-react-native/interactive-hooks to v0.21.1
+- Bump @fluentui-react-native/pressable to v0.9.34
+- Bump @fluentui-react-native/text to v0.16.1
+- Bump @uifabricshared/foundation-compose to v1.12.21
+
+## 0.17.3
+
+Sun, 13 Nov 2022 09:17:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.21.0
+- Bump @fluentui-react-native/pressable to v0.9.33
+- Bump @fluentui-react-native/text to v0.16.0
+
+## 0.17.2
+
+Fri, 11 Nov 2022 18:06:31 GMT
+
+### Patches
+
+- Bump all design-tokens packages to 0.29.0 (78454019+lyzhan7@users.noreply.github.com)
+- Bump @fluentui-react-native/framework to v0.8.19
+- Bump @fluentui-react-native/interactive-hooks to v0.20.2
+- Bump @fluentui-react-native/pressable to v0.9.32
+- Bump @fluentui-react-native/text to v0.15.15
+- Bump @fluentui-react-native/theme-tokens to v0.21.3
+- Bump @uifabricshared/foundation-compose to v1.12.20
 
 ## 0.17.1