+++ packages/experimental/Checkbox/CHANGELOG.md
@@ -1,8 +1,34 @@
 # Change Log - @fluentui-react-native/experimental-checkbox
 
-This log was last generated on Wed, 09 Nov 2022 22:52:11 GMT and should not be manually modified.
+This log was last generated on Wed, 16 Nov 2022 08:47:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.4
+
+Wed, 16 Nov 2022 08:47:25 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.17.4
+- Bump @fluentui-react-native/framework to v0.8.20
+
+## 0.13.3
+
+Sun, 13 Nov 2022 09:17:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.17.3
+
+## 0.13.2
+
+Fri, 11 Nov 2022 18:06:32 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/checkbox to v0.17.2
+- Bump @fluentui-react-native/framework to v0.8.19
 
 ## 0.13.1